+++ demo/assembly/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 # DataSource configuration
 spring.datasource.url=jdbc:postgresql://localhost:5432/demo
 spring.datasource.username=postgres
-spring.datasource.password=1234
+spring.datasource.password=Din kode
 spring.datasource.driver-class-name=org.postgresql.Driver
 
 # JPA (Java Persistence API) configuration
