+++ demo/main-app/pom.xml
@@ -39,6 +39,11 @@
             <artifactId>assembly</artifactId>
             <version>0.0.1-SNAPSHOT</version>
         </dependency>
+        <dependency>
+            <groupId>com.example</groupId>
+            <artifactId>product</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+        </dependency>
     </dependencies>
 
     <!-- Repository declarations and build configuration here -->
