+++ demo/product/src/main/java/com/example/product/BatchCreator.java
@@ -1,0 +1,8 @@
+package com.example.product;
+
+
+
+
+public class BatchCreator {
+
+}
