+++ demo/product/src/main/java/com/example/product/dto/PartDTO.java
@@ -1,0 +1,25 @@
+package com.example.product.dto;
+
+
+// Data Transfer Object (DTO)
+public class PartDTO {
+    private String name;
+    private String description;
+
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+}
