+++ demo/product/src/main/java/com/example/product/model/Product.java
@@ -1,0 +1,58 @@
+package com.example.product.model;
+
+import jakarta.persistence.*;
+
+import java.util.List;
+
+@Entity
+public class Product {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String name;
+    @OneToOne(optional = true)
+    private Recipe recipe;
+    @Column(nullable = true,length = 1_000)
+    private String description;
+
+    public Product() {
+    }
+
+    public Product(String name, Recipe recipe) {
+        this.name = name;
+        this.recipe = recipe;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Recipe getRecipe() {
+        return recipe;
+    }
+
+    public void setRecipe(Recipe recipe) {
+        this.recipe = recipe;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    @Override
+    public String toString() {
+        return "Product{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", recipe=" + recipe +
+                '}';
+    }
+}
