+++ demo/product/src/main/java/com/example/product/repository/BatchRepository.java
@@ -1,0 +1,7 @@
+package com.example.product.repository;
+
+import com.example.product.model.Batch;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BatchRepository extends JpaRepository<Batch,Long> {
+}
