+++ demo/product/src/main/java/com/example/product/repository/PartRepository.java
@@ -1,0 +1,11 @@
+package com.example.product.repository;
+
+import com.example.product.model.Part;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface PartRepository extends JpaRepository<Part,Long> {
+    Optional<Part> findByName(String name);
+}
+
