+++ demo/product/src/main/java/com/example/product/repository/ProductRepository.java
@@ -1,0 +1,7 @@
+package com.example.product.repository;
+
+import com.example.product.model.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product,Long> {
+}
