+++ demo/product/src/main/java/com/example/product/repository/RecipeRepository.java
@@ -1,0 +1,7 @@
+package com.example.product.repository;
+
+import com.example.product.model.Recipe;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RecipeRepository extends JpaRepository<Recipe,Long> {
+}
