+++ demo/product/src/main/java/module-info.java
@@ -4,6 +4,11 @@ module product.module {
     requires spring.web;
     requires spring.data.jpa;
     requires com.fasterxml.jackson.databind;
+    requires spring.context;
+    requires spring.beans;
+    requires spring.tx;
+    requires jakarta.persistence;
 
     exports com.example.product;
+
 }
