--- demo/src/main/java/com/example/demo/products/dto/ComponentDTO.java
@@ -1,25 +1,0 @@
-package com.example.demo.products.dto;
-
-
-// Data Transfer Object (DTO)
-public class ComponentDTO {
-    private String name;
-    private String description;
-
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-}
