--- demo/src/main/java/com/example/demo/products/model/Batch.java
@@ -1,49 +1,0 @@
-package com.example.demo.products.model;
-
-
-import jakarta.persistence.*;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Entity
-public class Batch {
-
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @OneToMany
-    private List<Product> products;
-    private LocalDateTime createdAt;
-    private boolean completed;
-
-    public Batch() {
-        this.createdAt=LocalDateTime.now();
-    }
-
-    public List<Product> getProducts() {
-        return products;
-    }
-
-    public void setProducts(List<Product> products) {
-        this.products = products;
-    }
-
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(LocalDateTime createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public boolean isCompleted() {
-        return completed;
-    }
-
-    public void setCompleted(boolean completed) {
-        this.completed = completed;
-    }
-}
