--- demo/src/main/java/com/example/demo/products/model/RecipeComponent.java
@@ -1,48 +1,0 @@
-package com.example.demo.products.model;
-
-import jakarta.persistence.*;
-
-@Entity
-public class RecipeComponent {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "recipe_id")
-    private Recipe recipe;
-
-    @ManyToOne
-    @JoinColumn(name = "component_id")
-    private Component component;
-
-    private int quantity;
-
-    public RecipeComponent() {
-    }
-
-    public Recipe getRecipe() {
-        return recipe;
-    }
-
-    public void setRecipe(Recipe recipe) {
-        this.recipe = recipe;
-    }
-
-    public Component getComponent() {
-        return component;
-    }
-
-    public void setComponent(Component component) {
-        this.component = component;
-    }
-
-    public int getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(int quantity) {
-        this.quantity = quantity;
-    }
-}
