--- demo/src/main/java/com/example/demo/products/repository/BatchRepository.java
@@ -1,7 +1,0 @@
-package com.example.demo.products.repository;
-
-import com.example.demo.products.model.Batch;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BatchRepository extends JpaRepository<Batch,Long> {
-}
