--- demo/src/main/java/com/example/demo/products/repository/ProductRepository.java
@@ -1,7 +1,0 @@
-package com.example.demo.products.repository;
-
-import com.example.demo.products.model.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product,Long> {
-}
