--- demo/src/main/java/com/example/demo/products/repository/RecipeRepository.java
@@ -1,7 +1,0 @@
-package com.example.demo.products.repository;
-
-import com.example.demo.products.model.Recipe;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RecipeRepository extends JpaRepository<Recipe,Long> {
-}
