--- demo/src/main/java/com/example/demo/products/service/BatchCreator.java
@@ -1,8 +1,0 @@
-package com.example.demo.products.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class BatchCreator {
-
-}
