--- demo/src/main/resources/application.properties
@@ -1,14 +1,0 @@
-# DataSource configuration
-spring.datasource.url=jdbc:postgresql://localhost:5432/demo
-spring.datasource.username=postgres
-spring.datasource.password=1234
-spring.datasource.driver-class-name=org.postgresql.Driver
-
-# JPA (Java Persistence API) configuration
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.hibernate.ddl-auto=update
-#spring.jpa.hibernate.ddl-auto=create-drop
-spring.jpa.show-sql=true
-
-logging.level.org.springframework.ws=DEBUG
-
