--- demo/src/test/java/com/example/demo/DemoApplicationTests.java
@@ -1,13 +1,0 @@
-package src.test.java.com.example.demo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
